test: Move the function responsible for decoding a YAML MeteringConfig manifest to Go to the testhelpers.
